// frontend/src/lib/api/FilterType.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type FilterType =
  | 'luminance'
  | 'red'
  | 'green'
  | 'blue'
  | 'h_alpha'
  | 'oiii'
  | 'sii'
  | 'uv_ir_cut'
  | 'dual_band'
  | 'tri_band'
  | 'quad_band'
  | 'light_pollution'
  | 'broadband_color'
  | 'other';